LayoutTests/imported/w3c/web-platform-tests/css/cssom/font-family-serialization-001-expected.txt


PASS Serialization of <generic-family>
FAIL Serialization of quoted "<generic-family>" assert_equals: expected "\"serif\"" but got "serif"
FAIL Serialization of prefixed -webkit-<generic-family> assert_equals: expected "-webkit-serif" but got "serif"
FAIL Serialization of NonGenericFontFamilyName,-webkit-body,-webkit-standard,-webkit-pictograph,BlinkMacSystemFont assert_equals: expected "-webkit-body" but got "Times"
